Just the names on a search result
The thumbnail and the category chip come off the row: what a text
search returns is a sentence somebody wrote, and three pieces of
furniture in front of it were competing with the quote for the eye.
Scientific name, common name in brackets, the section it matched.

SpeciesThumbnail goes back into NearbySpeciesPanel where it came from,
since nothing else uses it now and a shared component with one caller
is just a file to go and find. CitationProse stays shared — the search
does use it. The panel's diff against main is that one import and its
one call site.

The search's own query drops two columns with them: the species lookup
is common names alone.
